FIX XSS vulnerability in options page
When the user enters banner text in the options page
they are able to use HTML tags to format their banner.

We already sanitise this input with wp_kses_post which
allows "safe" HTML tags, but not unsafe ones such as
"script".

However, the banner text is passed to wp_kses_post()
as HTML encoded form data from the client, which means
that some characters have already been escaped. So
that wp_kses_post() can correctly sanitise the
user input, we need to remove this with html_entity_decode().
Without the call to html_entity_decode() unsafe
tags such as "script" are not correctly removed.